fix: 마이페이지 상단바 수정
