+++ PayKids/presentation/src/main/java/com/paykids/presentation/view/home/HomeFragment.kt
@@ -54,6 +54,7 @@ class HomeFragment : BaseFragment<FragmentHomeBinding>() {
         }
 
         hideStatusBar()
+        setStatusBarColorLight()
     }
 
     override fun initListener() {
