+++ PayKids/presentation/src/main/java/com/paykids/presentation/view/quest/AchievementFragment.kt
@@ -48,4 +48,9 @@ class AchievementFragment : BaseFragment<FragmentAchievementBinding>() {
             }
         }
     }
+
+    override fun onResume() {
+        super.onResume()
+        setStatusBarColorDark()
+    }
 }
