+++ PayKids/presentation/src/main/res/layout/fragment_modify_info.xml
@@ -20,12 +20,14 @@
             android:fontFamily="@font/nanumsquare_extra_bold"
             android:gravity="center"
             android:includeFontPadding="false"
+            android:paddingTop="40dp"
 
             android:text="@string/text_my_page"
             android:textColor="@color/black"
             android:textSize="@dimen/text_size_header"
 
             app:layout_constraintEnd_toEndOf="parent"
+            app:layout_constraintHorizontal_bias="0.0"
             app:layout_constraintStart_toStartOf="parent"
             app:layout_constraintTop_toTopOf="parent" />
 
@@ -33,6 +35,7 @@
             android:id="@+id/ib_back"
             android:layout_width="48dp"
             android:layout_height="48dp"
+            android:layout_marginTop="36dp"
             android:background="@color/transparent"
             android:contentDescription="@null"
             android:src="@drawable/ic_left_blue"
