+++ PayKids/presentation/src/main/res/layout/fragment_mypage.xml
@@ -15,6 +15,7 @@
         android:gravity="center"
         android:includeFontPadding="false"
 
+        android:paddingTop="40dp"
         android:text="@string/text_my_page"
         android:textColor="@color/black"
         android:textSize="@dimen/text_size_header"
