+++ PayKids/presentation/src/main/res/values/dimens.xml
@@ -3,7 +3,7 @@
     <dimen name="button_height">54dp</dimen>
     <dimen name="edittext_height">48dp</dimen>
     
-    <dimen name="header_height">60dp</dimen>
+    <dimen name="header_height">100dp</dimen>
     <dimen name="text_size_header">20sp</dimen>
 
     <dimen name="grid_spacing">8dp</dimen>
